requirements.txt
# core HTTP & async
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.2
async-lru==2.0.4

# general utilities
appdirs==1.4.4
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
filelock==3.12.1
packaging==23.1
platformdirs==3.10.0
pyyaml==6.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.16
wasabi==1.1.2
webencodings==0.5.1

# HTML/XML/Text
beautifulsoup4==4.12.2
lxml==4.9.3
cssselect==1.1.0
markdown==3.8
Pygments==2.16.1

# Data & ML
numpy==1.24.4
pandas==2.0.2
scikit-learn==1.3.2
scipy==1.11.1
h5py==3.9.0
pyarrow==12.0.0
tables==3.8.0

# Image I/O / Vision
Pillow==10.0.0
opencv-python-headless==4.8.1.78
tifffile==2024.3.1
imageio==2.27.0

# Deep learning frameworks
torch==2.0.1+cu118           # or cpu-only: torch==2.0.1
torchvision==0.15.2+cu118    # match your torch build
torchaudio==2.0.2+cu118      # if needed
timm==0.9.2

# Transformers & LLMs
transformers
tokenizers==0.13.3
bitsandbytes==0.39.1         # for 8‑bit LLM training
accelerate==0.24.0
datasets==2.12.0
huggingface-hub==0.14.1
sentencepiece==0.1.99

# Report cleaning & metrics
bert-score==0.3.13
fast-bleu==0.0.90
seqeval==1.2.2
nlpaug==1.1.10               # optional augmentations

# Retrieval & embeddings
faiss-cpu==1.7.4             # or faiss-gpu==1.7.4 for GPU
scikit-learn==1.3.2          # already above, for vector search

# Logging & tracking
wandb==0.15.5
yaml==6.0

# Visualization (optional)
matplotlib==3.8.0
seaborn==0.12.2
# radgraph @ git+https://github.com/stanfordnlp/radgraph.git@main

# Small utility libraries
psutil==5.9.5
pyyaml==6.0

# Additional libraries:
f1chexbert
bert-score
huggingface_hub
peft
datasets
deepspeed
openai==0.26.0
bitsandbytes
accelerate
